s3-tests-rs: add admin module and quota user support
Move admin_request() into the shared library so all test binaries
can use it.  Add driver_hint() convenience wrapper for the new
driver hint admin endpoint.

Add [s3 quota] config section and get_quota_client() for a dedicated
quota test user, isolated from the main test user.  Update cleanup
to nuke quota user buckets when configured.

Update s3tests.conf.SAMPLE with the new section.
