+++ src/test/rgw/s3-tests-rs/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod admin;
 pub mod cleanup;
 pub mod client;
 pub mod config;
